fix: code cleanup -- better in-code documentation and descriptive naming conventions, use NPE checks in appropriate places, and remove unused scss.
